move umask_context to tests/util.py module; fix NN ensemble LMDB group write permissions and verify using test
